# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aiscast API",
    "version": "1.0.0",
    "description": "Live AIS data from openwaters.io: current vessel positions, receiving stations, usage stats, self-serve tokens, and HTTP ingest for feeders. Successful responses are JSON; errors and `/health` are plain text. CORS is open, times are UTC, and JSON is compressed (zstd or gzip) when the request sends `Accept-Encoding`.\n\nThe streams are documented separately: `/v0/stream` (WebSocket, aisstream.io-compatible), `/v1/stream` (WebSocket JSON frames, MQTT over WebSocket for feeders, or Server-Sent Events), `/v1/nmea` (WebSocket, raw NMEA), and raw NMEA ingest on UDP port 10110. `/mcp` is the MCP (Model Context Protocol) endpoint for AI assistants: Streamable HTTP, no sign-in, the same tiers as `/v1/vessels`. See [openwaters.io/api/ais](https://openwaters.io/api/ais/).",
    "contact": {
      "name": "aiscast",
      "url": "https://github.com/openwatersio/aiscast"
    },
    "termsOfService": "https://github.com/openwatersio/aiscast/blob/main/docs/contributor-agreement.md",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [{ "url": "https://ais.openwaters.io" }],
...
  "paths": {
...
    "/v1/vessels/tiles/{z}/{x}/{y}": {
      "get": {
        "operationId": "getVesselTile",
        "tags": ["Vessels and stations"],
        "summary": "Vessel tile by z/x/y",
        "description": "One vector tile, at the URL template that `/v1/vessels/tiles.json` points to. Most maps should use the TileJSON instead. Use this template directly only for a client that takes a tile URL rather than TileJSON. It takes the same parameters and returns the same `vessels` layer. Tiles have their own rate limit of 600 requests per minute per address.",
        "security": [{}, { "token": [] }],
        "parameters": [
          { "name": "z", "in": "path", "required": true, "description": "Zoom, 0–22. TileJSON advertises 14 as the maximum, and clients overzoom past it.", "schema": { "type": "integer", "minimum": 0, "maximum": 22 } },
          { "name": "x", "in": "path", "required": true, "description": "Column, 0 to 2^z−1.", "schema": { "type": "integer", "minimum": 0 } },
          { "name": "y", "in": "path", "required": true, "description": "Row, 0 to 2^z−1, from the north.", "schema": { "type": "integer", "minimum": 0 } },
          { "$ref": "#/components/parameters/tileMMSI" },
          { "$ref": "#/components/parameters/kind" },
          { "$ref": "#/components/parameters/class" },
          { "$ref": "#/components/parameters/type" },
          { "$ref": "#/components/parameters/min_sog" },
          { "$ref": "#/components/parameters/max_age_moving" },
          { "$ref": "#/components/parameters/tileMaxAge" }
        ],
        "responses": {
          "200": {
            "description": "The tile, gzipped when the request accepts gzip. An empty tile has an empty `vessels` layer.",
            "content": { "application/vnd.mapbox-vector-tile": { "schema": { "type": "string", "format": "binary" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An aiscast claim token (`ak1....`), created by `POST /v1/keys` or issued by the operator. Besides `Authorization: Bearer`, the token is accepted as an HTTP Basic password (`anything:<token>`, which is what AIS-catcher's USERPWD sends) or as `?key=<token>` on the URL. A token that is present but invalid, expired, revoked, used outside its `cidr`, or lacking the role for the action is refused; it never degrades to anonymous access."
      }
    },
    "parameters": {
...
      "tileMMSI": { "name": "mmsi", "in": "query", "required": false, "description": "Comma-separated MMSIs: only these vessels. The MMSI cap applies, as on `/v1/vessels`.", "schema": { "type": "string" }, "example": "368168720,257090090" },
      "kind": { "name": "kind", "in": "query", "required": false, "description": "Comma-separated: `vessel`, `aton` (aids to navigation), `base` (base stations), `sar` (search and rescue aircraft).", "schema": { "type": "string" }, "example": "vessel,sar" },
      "class": { "name": "class", "in": "query", "required": false, "description": "`A` or `B`, the transponder class from the position reports. A vessel whose class is not yet known is left out.", "schema": { "type": "string", "enum": ["A", "B"] } },
      "type": { "name": "type", "in": "query", "required": false, "description": "Comma-separated ITU ship and cargo type codes or ranges, such as `70-89` for cargo and tankers or `36,37` for sailing and pleasure craft. For an aid to navigation, the AtoN type. A vessel whose type is not yet known is type 0.", "schema": { "type": "string" }, "example": "70-89" },
      "min_sog": { "name": "min_sog", "in": "query", "required": false, "description": "Only vessels at least this fast, in knots. A vessel with no known speed is left out.", "schema": { "type": "number", "minimum": 0 }, "example": 0.5 },
      "tileMaxAge": { "name": "max_age", "in": "query", "required": false, "description": "Oldest last report to include: whole seconds, a duration such as `24h`, or `all`. Defaults to 7 days, or to `all` when `mmsi` is given.", "schema": { "type": "string" }, "example": "24h" },
      "max_age_moving": { "name": "max_age_moving", "in": "query", "required": false, "description": "Oldest last report to include for a vessel whose last report was under way, that is, not stationary as `/v1/vessels` defines it: whole seconds, a duration such as `2h`, or `all`. Defaults to 30 minutes for an area and to `all` for vessels named in `mmsi`.", "schema": { "type": "string" }, "example": "all" }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed parameter or body. The plain-text body says which.",
        "content": { "text/plain": { "schema": { "type": "string" } } }
      },
      "Unauthorized": {
        "description": "Missing, invalid, expired, or revoked token, or the token lacks the role for this action. The plain-text body gives the reason.",
        "content": { "text/plain": { "schema": { "type": "string" } } }
      },
      "RateLimited": {
        "description": "Over the per-address or per-station request limit.",
        "content": { "text/plain": { "schema": { "type": "string" }, "example": "rate limited" } }
      }
    },
...
  }
}